fix audit

+++ kubemarine/procedures/install.py
@@ -141,11 +141,11 @@ def system_prepare_policy(cluster):
             master.sudo("kubeadm init phase control-plane apiserver --config=/etc/kubernetes/audit-on-config.yaml")
             master.sudo("kubeadm init phase upload-config kubeadm --config=/etc/kubernetes/audit-on-config.yaml")
 
-
-    cluster.nodes['master'].get_first_member().call(utils.wait_command_successful, command="crictl rm -f "
-                                                                                           "$(sudo crictl ps | grep 'kube-apiserver'"
-                                                                                           " | awk '{ print $1 }')")
-    cluster.nodes['master'].get_first_member().call(utils.wait_command_successful, command="kubectl get pod -A")
+    if kubernetes.is_cluster_installed(cluster):
+        cluster.nodes['master'].get_first_member().call(utils.wait_command_successful, command="crictl rm -f "
+                                                                                               "$(sudo crictl ps | grep 'kube-apiserver'"
+                                                                                             " | awk '{ print $1 }')")
+        cluster.nodes['master'].get_first_member().call(utils.wait_command_successful, command="kubectl get pod -A")
 
 def system_prepare_dns_hostname(cluster):
     with RemoteExecutor(cluster):
